Rename infra/database into infra/orm
