+++ lib/interfaces/storage/UserRepositorySQLite.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const sequelize = require('../../infrastructure/database/sequelize/sequelize');
+const sequelize = require('../../infrastructure/orm/sequelize/sequelize');
 const User = require('../../domain/User');
 const UserRepository = require('../../application/repositories/UserRepository');
